net/igc/base: fix bitwise operation type
[ upstream commit bf51125b75b18a3a43439edb2c6e1f0b14793d51 ]

Static analysis has found type mismatch between mask and register value.
Fix the bitwise operation type to avoid potential issues.

Fixes: 8cb7c57d9b3c ("net/igc: support device initialization")

Acked-by: Bruce Richardson <bruce.richardson@intel.com>

+++ drivers/net/igc/base/igc_i225.c
@@ -341,7 +341,7 @@ void igc_release_swfw_sync_i225(struct igc_hw *hw, u16 mask)
 	}
 
 	swfw_sync = IGC_READ_REG(hw, IGC_SW_FW_SYNC);
-	swfw_sync &= ~mask;
+	swfw_sync &= ~(u32)mask;
 	IGC_WRITE_REG(hw, IGC_SW_FW_SYNC, swfw_sync);
 
 	igc_put_hw_semaphore_generic(hw);
